footer

+++ VliegWebsite/Delete.php
@@ -1,22 +1,18 @@
 <?php
-
 include('connection.php');
 
-
 if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['delete'])) {
+    if (isset($_POST['vluchtID'])) {
+        $vluchtID = $_POST['vluchtID'];
+        $sql = "DELETE FROM vlucht WHERE vluchtID = :vluchtID";
+        $stmt = $conn->prepare($sql);
+        $stmt->bindParam(':vluchtID', $vluchtID);
+        if ($stmt->execute()) {
+            echo "Product succesvol verwijderd";
+        } else {
+            echo "Error deleting product: " . $stmt->errorInfo()[2];
+        }
 
-    $sql = "DELETE FROM vlucht WHERE vluchtID = :vluchtID";
-    $stmt = $conn->prepare($pdo);
-    $stmt->bindParam(':vluchtID', $vluchtID);
-
-    if ($stmt->execute()) {
-        echo "Product verwijderd succesvol";
-    } else {
-        echo "Error deleting product: " . $stmt->errorInfo()[2]; 
-    }
 }
 
-$stmt = $conn->query("SELECT * FROM vlucht");
-$vlucht = $stmt->fetchAll();
-
-?>
+}
